+++ AdapterUnitTests/VerveAdapterTests/VerveAdapterTest.swift
@@ -8,6 +8,9 @@ final class VerveAdapterTest: XCTestCase {
 
   override func tearDown() {
     HybidClientFactory.debugClient = nil
+    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = nil
+    MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = nil
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .unspecified
   }
 
   func testAdapterVersion() {
@@ -68,9 +71,21 @@ final class VerveAdapterTest: XCTestCase {
       VerveAdapterError(errorCode: .childUser, description: "some error message").toNSError())
   }
 
+  func testAdapterSetUp_fails_whenAgeRestrictedTreatmentIsChild() {
+    HybidClientFactory.debugClient = nil
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .child
+
+    let credentials = AUTKMediationCredentials()
+    credentials.settings = ["AppToken": "AppToken"]
+    let serverConfiguration = AUTKMediationServerConfiguration()
+    serverConfiguration.credentials = [credentials]
+    AUTKWaitAndAssertAdapterSetUpFailureWithConfiguration(
+      VerveAdapter.self, serverConfiguration,
+      VerveAdapterError(errorCode: .childUser, description: "some error message").toNSError())
+  }
+
   func testAdapterSetUp_fails_whenAppTokenIsMissing() {
     HybidClientFactory.debugClient = nil
-    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = true
 
     let credentials = AUTKMediationCredentials()
     let serverConfiguration = AUTKMediationServerConfiguration()
@@ -260,6 +275,296 @@ final class VerveAdapterTest: XCTestCase {
       XCTAssertNotNil(signals)
       XCTAssertNil(error)
       expectation.fulfill()
+    }
+    wait(for: [expectation])
+  }
+
+  func testCollectionSignals_fails_whenCOPPAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "collect signals")
+    let params = AUTKRTBRequestParameters()
+
+    MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = true
+
+    adapter.collectSignals(for: params) { signals, error in
+      XCTAssertEqual(signals, "")
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+    }
+    wait(for: [expectation])
+  }
+
+  func testCollectionSignals_fails_whenTFUAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "collect signals")
+    let params = AUTKRTBRequestParameters()
+
+    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = true
+
+    adapter.collectSignals(for: params) { signals, error in
+      XCTAssertEqual(signals, "")
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+    }
+    wait(for: [expectation])
+  }
+
+  func testCollectionSignals_fails_whenAgeRestrictedTreatmentIsChild() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "collect signals")
+    let params = AUTKRTBRequestParameters()
+
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .child
+
+    adapter.collectSignals(for: params) { signals, error in
+      XCTAssertEqual(signals, "")
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+    }
+    wait(for: [expectation])
+  }
+
+  // MARK: - loadBanner Tests under isChildUser
+
+  func testLoadBanner_fails_whenCOPPAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load banner")
+    let config = AUTKMediationBannerAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = true
+
+    adapter.loadBanner(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadBanner_fails_whenTFUAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load banner")
+    let config = AUTKMediationBannerAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = true
+
+    adapter.loadBanner(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadBanner_fails_whenAgeRestrictedTreatmentIsChild() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load banner")
+    let config = AUTKMediationBannerAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .child
+
+    adapter.loadBanner(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  // MARK: - loadInterstitial Tests under isChildUser
+
+  func testLoadInterstitial_fails_whenCOPPAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load interstitial")
+    let config = AUTKMediationInterstitialAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = true
+
+    adapter.loadInterstitial(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadInterstitial_fails_whenTFUAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load interstitial")
+    let config = AUTKMediationInterstitialAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = true
+
+    adapter.loadInterstitial(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadInterstitial_fails_whenAgeRestrictedTreatmentIsChild() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load interstitial")
+    let config = AUTKMediationInterstitialAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .child
+
+    adapter.loadInterstitial(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  // MARK: - loadRewardedAd Tests under isChildUser
+
+  func testLoadRewardedAd_fails_whenCOPPAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load rewarded")
+    let config = AUTKMediationRewardedAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = true
+
+    adapter.loadRewardedAd(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadRewardedAd_fails_whenTFUAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load rewarded")
+    let config = AUTKMediationRewardedAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = true
+
+    adapter.loadRewardedAd(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadRewardedAd_fails_whenAgeRestrictedTreatmentIsChild() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load rewarded")
+    let config = AUTKMediationRewardedAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .child
+
+    adapter.loadRewardedAd(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  // MARK: - loadNativeAd Tests under isChildUser
+
+  func testLoadNativeAd_fails_whenCOPPAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load native")
+    let config = AUTKMediationNativeAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = true
+
+    adapter.loadNativeAd(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadNativeAd_fails_whenTFUAIsTrue() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load native")
+    let config = AUTKMediationNativeAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent = true
+
+    adapter.loadNativeAd(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
+    }
+    wait(for: [expectation])
+  }
+
+  func testLoadNativeAd_fails_whenAgeRestrictedTreatmentIsChild() {
+    let adapter = VerveAdapter()
+    let expectation = expectation(description: "load native")
+    let config = AUTKMediationNativeAdConfiguration()
+
+    MobileAds.shared.requestConfiguration.ageRestrictedTreatment = .child
+
+    adapter.loadNativeAd(for: config) { ad, error in
+      XCTAssertNil(ad)
+      let nsError = error as? NSError
+      XCTAssertNotNil(nsError)
+      XCTAssertEqual(nsError?.domain, "com.google.mediation.verve")
+      XCTAssertEqual(nsError?.code, 102)
+      expectation.fulfill()
+      return nil
     }
     wait(for: [expectation])
   }